Add support for indexed type vars and env keys in compiler
